Avoiding to send the cookie header on every request
The default configuration makes the cookie header being set on every request. Besides increasing the bandwidth, make the browser to reprocess the cookies. Also, and the most important fact, it also block caching tools to cache the page (if allowed). Most of the tools ignore the caching if any cache is being set.