Fixed false positive for diverged
If the configuration has extra information that is ignored by the Swivel core, it will still tag as diverged when it is not. Ie, if the feature is configured for the buckets 1,2,3,4,5,6,7,8,9,10,11 it is all on. The 11 is ignored by Swivel, so it should be ignored when checking the divergency too. Same happens with bucket 0 that can be appear in some cases.